Refactor manage seat price
