--- Jolly-Lights-Cinema-Group/BusinessLogic/GlobalHandlers/SeatHandler.cs
@@ -1,10 +1,0 @@
-using Jolly_Lights_Cinema_Group.Common;
-using Jolly_Lights_Cinema_Group.Models;
-using JollyLightsCinemaGroup.DataAccess;
-
-namespace Jolly_Lights_Cinema_Group.BusinessLogic.GlobalHandlers;
-
-public class SeatHandler
-{
-
-}
